requirements.txt
langchain==0.1.16
langchain-openai==0.1.3
langchain-community==0.0.32
openai>=1.30.0
faiss-cpu
streamlit
python-dotenv
pypdf
